separate logging funcs from main.rs. write test also
